requirements.txt
python-telegram-bot==13.15
redis==5.0.0
python-dotenv==1.0.0
requests==2.31.0
